% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_summary_table_helpers.R
\name{add_row}
\alias{add_row}
\title{Update warning text data frame}
\usage{
add_row(df, data_check, topic, group, analysis)
}
\arguments{
\item{df}{Data frame with existing warning texts}

\item{data_check}{Output from data_available()}

\item{topic}{Safety topics}

\item{group}{Comparison number}

\item{analysis}{Type of analysis}
}
\value{
Another row for a dataframe returning warning texts about missing information and its consequence.
}
\description{
Adds another row, if data is missing in a comparison
}
