% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{app_plots}
\alias{app_plots}
\title{App plots}
\usage{
app_plots(
  select_analysis,
  map_object,
  saf_topic,
  select_btrt,
  new_trial_analysis,
  stat_inf_dist,
  mix
)
}
\arguments{
\item{select_analysis}{event propotions or rates}

\item{map_object}{MAP Prior object}

\item{saf_topic}{Selected safety topic to analyze/the adverse event of interest}

\item{select_btrt}{selected treatment arm}

\item{new_trial_analysis}{nta}

\item{stat_inf_dist}{Data frame with statistical information of the distribution}

\item{mix}{mixture distribution object}
}
\description{
Only includes plots generated in the app
}
