% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_summary_table_helpers.R
\name{array_comp}
\alias{array_comp}
\title{Ary Comparison}
\usage{
array_comp(
  ana,
  array_comp = NA,
  comp,
  pr_sample,
  topic = topic,
  group,
  crilb = lb,
  criub = ub,
  array_ana = NA
)
}
\arguments{
\item{array_comp}{The array setup for comparisons}

\item{comp}{"Risk Diff" or "Risk Ratio"}

\item{pr_sample}{posterior sample for difference or ratio}

\item{topic}{Variable to be analyzed}

\item{group}{The current comparison group}

\item{crilb}{Cri lower bound}

\item{criub}{Cri upper bound}

\item{array_ana}{Setup for analysed array}

\item{sample_size}{sample size}
}
\description{
Assigning values from a sample to array_inci_comp
}
