% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_summary_table_helpers.R
\name{array_rmix}
\alias{array_rmix}
\title{Array rmix}
\usage{
array_rmix(rmix_obj, array, arm, topic, group, lb = lb, ub = ub)
}
\arguments{
\item{rmix_obj}{A S3 class matrix from R RBesT with a mixture distribution}

\item{array}{The array setup}

\item{arm}{"Arm A" (treatment) or "Arm B" (control)}

\item{topic}{Safety Topic}

\item{group}{Comparison number}

\item{lb}{Lower bound for credible interval}

\item{ub}{Upper bound for credible interval}
}
\description{
Assigning values from a rmix object from RBesT to the array_inci
}
