% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_helpers.R
\name{check_minmax}
\alias{check_minmax}
\title{Beta Dist Update check}
\usage{
check_minmax(r, n)
}
\arguments{
\item{r}{Response = alpha}

\item{n}{Total = alpha + beta}
}
\value{
A vector containing alpha and beta for the beta distribution
}
\description{
Many unnecessary error occur with RBesT when in a Beta Dist
the parameters are <1
}
