% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_helpers.R
\name{cri_char}
\alias{cri_char}
\title{Helper for credible interval display}
\usage{
cri_char(df = NA)
}
\arguments{
\item{df}{A dataframe that is expected to include the columns crilb and criub}
}
\value{
a data frame including the credible intervals
}
\description{
Put Credible Interval in one Character
}
