% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_table_prep}
\alias{data_table_prep}
\title{Summary Table Preparation}
\usage{
data_table_prep(
  input_data,
  select_analysis,
  saf_topic,
  select_btrt,
  bool_pooled = FALSE,
  current_trial = FALSE
)
}
\arguments{
\item{input_data}{the raw summary level data}

\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}

\item{saf_topic}{Selected safety topic to analyze/the adverse event of interest}

\item{select_btrt}{selected background treatment}

\item{bool_pooled}{Whether study data is pooled by study number}

\item{current_trial}{information whether it is a historical (0) or current (1) trial}
}
\value{
a dataframe for the selected safety topic with the study ID,
the number of patients in the selected arm, the number of patients in the arm with at
least one occurrence of the adverse event, and the total exposure in patient years
if the exposure-adjusted analysis is chosen
}
\description{
Returns a dataframe for the selected safety topic with the study ID,
the number of patients in the selected arm, the number of patients in the arm with at
least one occurrence of the adverse event, and the total exposure in patient years
if the exposure-adjusted analysis is chosen.
}
