% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{decision_making_density_plot}
\alias{decision_making_density_plot}
\title{Title}
\usage{
decision_making_density_plot(
  select_analysis,
  stat_inf_dist,
  ae_prop,
  saf_topic,
  select_btrt
)
}
\arguments{
\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}

\item{stat_inf_dist}{Data frame with statistical information of the distribution}

\item{ae_prop}{Vector of quantiles}

\item{saf_topic}{Selected safety topic to analyze/the adverse event of interest}

\item{select_btrt}{}
}
\description{
Title
}
