% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_summary_table_helpers.R
\name{inci_naiv}
\alias{inci_naiv}
\title{Naive incidence}
\usage{
inci_naiv(data, array_inci, arm, topic, group)
}
\arguments{
\item{data}{Historic and Current data}

\item{array_inci}{The array setup for incidence proportions}

\item{arm}{"Arm A" (treatment) or "Arm B" (control)}

\item{topic}{Safety Topic}

\item{group}{Comparison number}
}
\description{
Gives the naiv estimation outputs for the whole data set
}
