% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{input_data_display}
\alias{input_data_display}
\title{Display input data}
\usage{
input_data_display(data, select_analysis, saf_topic)
}
\arguments{
\item{data}{Uploaded Data}

\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}

\item{saf_topic}{Selected safety topic to analyze/the adverse event of interest}
}
\description{
Display input data in the historical data tab
}
