% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{map_prior_func}
\alias{map_prior_func}
\title{Meta-Analytic-Predictive Analysis for Generalized Linear Models}
\usage{
map_prior_func(
  input_data,
  select_analysis,
  tau_dist,
  adj_tau,
  seed = NULL,
  testing = FALSE
)
}
\arguments{
\item{input_data}{dataframe from data_table_prep()}

\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}

\item{tau_dist}{assumed distribution of tau}

\item{adj_tau}{numeric value from tau_adjust}

\item{seed}{a seed as input for reproducibility}

\item{testing}{for testing purposes faster MCMC}
}
\value{
an S3 object (list) of type gMAP for the incidence proportion analysis
or a vector of the posterior MCMC samples for the exposure-adjusted analysis
}
\description{
MAP prior calculation based on the selected analysis.
RBesT is used for the incidence proportion analysis.
MAP prior should be computed using only historical data
}
