% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_helpers.R
\name{mcmc_desc}
\alias{mcmc_desc}
\title{Describe MCMC sample}
\usage{
mcmc_desc(
  mcmc_obj = NA,
  crilb = 0.025,
  criub = 0.975,
  seed = NA,
  trans = FALSE
)
}
\arguments{
\item{mcmc_obj}{A mcmc object, or in general a sample vector}

\item{crilb}{Lower Bound for the credible interval}

\item{criub}{Upper Bound for the credible interval}

\item{seed}{Seed for reproducibility}

\item{trans}{If a Matrix with a mixture is given, it will be sampled from it.
Needed for interpretation of log(hazard)}
}
\value{
A data frame including descriptive statistics
}
\description{
Helper for descriptive statistics of sample
}
