% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{mix_distribution_all}
\alias{mix_distribution_all}
\title{Title}
\usage{
mix_distribution_all(
  select_analysis,
  current_trial_data,
  select_dist,
  param_approx,
  robust_map_object,
  post_dist
)
}
\arguments{
\item{post_dist}{}
}
\description{
Title
}
