% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{new_trial_compare}
\alias{new_trial_compare}
\title{Dataframe preparation for plotting robust map prior, likelihood, and posterior distributions}
\usage{
new_trial_compare(select_analysis, robust_map_prior, new_v1, new_v2, post_dist)
}
\arguments{
\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}

\item{robust_map_prior}{robust map prior; mixture distribution with a non-informative component from robust_map()}

\item{new_v1}{#new_n sample size in the new trial for the selected safety topic and background treatment}

\item{new_v2}{#new_r number of patients with at least one occurrence of the AE of interest}

\item{post_dist}{posterior mixture distribution or MCMC samples from posterior_dist()}
}
\value{
dataframe for plotting robust map prior, likelihood, and posterior distributions
}
\description{
Creates a dataframe for plotting robust map prior, likelihood, and posterior distributions
}
