% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{nta_data_conflict_assassment_plot}
\alias{nta_data_conflict_assassment_plot}
\title{Title}
\usage{
nta_data_conflict_assassment_plot(
  new_trial_analysis,
  saf_topic,
  select_btrt,
  select_analysis
)
}
\description{
Title
}
