% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{parametric_approx}
\alias{parametric_approx}
\title{Automatic Fitting of Mixtures of Conjugate Distributions to a Sample}
\usage{
parametric_approx(select_analysis, map_prior)
}
\arguments{
\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}

\item{map_prior}{an S3 object (list) of type gMAP for the incidence proportion analysis
or a vector of the posterior MCMC samples for the exposure-adjusted analysis}
}
\value{
the best fitting mixture model, i.e., the model with the lowest AIC
}
\description{
Fitting a series of mixtures of conjugate distributions to a sample
using Expectation-Maximization. First, a Nc\link{1} component mixture is fitted, then a
Nc\link{2} component mixture, and so on. The mixture providing the best AIC value is selected.
}
