% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{posterior_dist}
\alias{posterior_dist}
\title{Conjugate Posterior Analysis}
\usage{
posterior_dist(
  select_analysis,
  robust_map_prior,
  new_v1,
  new_v2,
  input_data,
  p_exch = NULL,
  nex_mean = NULL,
  nex_sd = NULL,
  adj_tau = NULL,
  seed = NULL
)
}
\arguments{
\item{select_analysis}{Incidence proportion and Exposure-adjusted AE rate}

\item{robust_map_prior}{mixture distribution output from robust_map()}

\item{new_v1}{first new variable #n for proportion and #hazard rate for rates}

\item{new_v2}{second new variable #r for proportion and unit information sd for rates}

\item{input_data}{dataframe from data_table_prep() including the new trial}

\item{p_exch}{probability of exchangeable component (use 1-robust_weight)}

\item{nex_mean}{Mean for Nex part (use robust_mean)}

\item{nex_sd}{Standard Deviation for Nex part}

\item{adj_tau}{numeric value of tau from tau_adjust()}

\item{seed}{for reproducibility}
}
\value{
posterior mixture distribution for incidence proportion analysis and
}
\description{
Calculates the posterior distribution for given prior robust_map_prior,
where the prior is a mixture of conjugate distributions. The posterior is then also
a mixture of conjugate distributions.
posterior MCMC samples from EXNEX analysis for exposure-adjusted analysis
}
