% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_helpers.R
\name{rmix_desc}
\alias{rmix_desc}
\title{Describe Mixture Distributions}
\usage{
rmix_desc(rmix_obj = NA, crilb = 0.025, criub = 0.975, deci = 4)
}
\arguments{
\item{rmix_obj}{A dataframe containing a mixture distribution}

\item{crilb}{Lower Bound for the credible interval}

\item{criub}{Upper Bound for the credible interval}

\item{deci}{Number of decimals for rounding}
}
\value{
A data frame including descriptive statistics
}
\description{
Helper for descriptive statistics of a matrix with mixture
distribution, as object from RBesT
}
