% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{robust_compare}
\alias{robust_compare}
\title{robust_compare}
\usage{
robust_compare(select_analysis, robust_map_prior, param_approx)
}
\arguments{
\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}

\item{robust_map_prior}{robust map prior; mixture distribution with a non-informative component from robust_map()}

\item{param_approx}{map prior; best fitting mixture model from parametric_approx()}
}
\value{
dataframe of comparison of map prior and robust map prior for ggplot
}
\description{
Creates the dataframe for plotting the comparison between the map prior and robust map prior
}
