% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{robust_map}
\alias{robust_map}
\title{Robustify Mixture Priors}
\usage{
robust_map(
  select_analysis,
  param_approx,
  input_data,
  robust_weight,
  robust_mean,
  adj_tau,
  seed = NULL,
  testing = FALSE
)
}
\arguments{
\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}

\item{param_approx}{the best fitting mixture model from parametric_approx()}

\item{input_data}{read in data}

\item{robust_weight}{weight given to the non-informative component (0 < weight < 1)}

\item{robust_mean}{mean of the non-informative component}

\item{adj_tau}{in between heterogeneity}

\item{seed}{seed for reprduceability}

\item{testing}{faster mcmc for testing}
}
\value{
new mixture distribution with an extra non-informative component
}
\description{
Add a non-informative component to a mixture prior
}
