% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{sampling_all_plot}
\alias{sampling_all_plot}
\title{Dataframe with density quantiles}
\usage{
sampling_all_plot(
  select_analysis,
  select_dist,
  param_approx,
  new_trial_analysis
)
}
\arguments{
\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}

\item{select_dist}{Choice of which should be displayed Robust MAP/MAP/Likelihood/Post}

\item{param_approx}{parametric approximation}

\item{new_trial_analysis}{Objects of the new trial analysis input}
}
\description{
MAP Prior, Robust MAP Prior, Likelihood, or Posterior Distribution Samples in the Decision Making Tab
}
