% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{tau_adjust}
\alias{tau_adjust}
\title{Tau Adjustment for Amount of Historical Borrowing}
\usage{
tau_adjust(select_analysis, hist_borrow)
}
\arguments{
\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}

\item{hist_borrow}{the amount of historical borrowing
("Small", "Moderate", "Substantial", "Large", "Very Large")}
}
\value{
the numeric value of tau based on the selected amount of historical borrowing
}
\description{
Change the value of tau to control the amount of historical borrowing
}
