% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_helpers.R
\name{text_prop}
\alias{text_prop}
\title{Helper for incidence proportion display}
\usage{
text_prop(stats_mat_prop = NA)
}
\arguments{
\item{stats_mat_prop}{Matrix with statiscial information for the proportional case}
}
\value{
Text for display
}
\description{
Data Frame to display proportions in \%
}
