% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{trt_data_wrangler}
\alias{trt_data_wrangler}
\title{Arm Choices}
\usage{
trt_data_wrangler(input_data)
}
\arguments{
\item{input_data}{the raw summary level data}
}
\description{
A function choose the Arms
}
