% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{app_plots}
\alias{app_plots}
\title{app_plots}
\usage{
app_plots(
  select_analysis,
  map_object,
  saf_topic,
  select_btrt,
  new_trial_analysis,
  stat_inf_dist,
  mix
)
}
\arguments{
\item{select_analysis}{event porpotions or rates}

\item{map_object}{Prior}

\item{saf_topic}{safety topic}

\item{select_btrt}{selected treatment arm}

\item{new_trial_analysis}{nta}

\item{stat_inf_dist}{differences}

\item{mix}{fitted mix}
}
\description{
Only includes plots generated in the app
}
