% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{decision_making_density_plot}
\alias{decision_making_density_plot}
\title{Title}
\usage{
decision_making_density_plot(
  select_analysis,
  stat_inf_dist,
  ae_prop,
  saf_topic,
  select_btrt
)
}
\arguments{
\item{select_btrt}{}
}
\description{
Title
}
