% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{forest_plot_display}
\alias{forest_plot_display}
\title{Title}
\usage{
forest_plot_display(map_object, select_analysis, saf_topic, select_btrt)
}
\arguments{
\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}

\item{saf_topic}{Selected safety topic to analyze/the adverse event of interest}
}
\description{
Title
}
