% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{map_prior_function_display}
\alias{map_prior_function_display}
\title{Parametric Approximation density display}
\usage{
map_prior_function_display(param_approx, select_analysis)
}
\arguments{
\item{param_approx}{parametric approximation}

\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}
}
\description{
Display parametric approximation mixture density function in the MAP Prior Tab
}
