% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{nta_data_conflict_assassment_plot}
\alias{nta_data_conflict_assassment_plot}
\title{Title}
\usage{
nta_data_conflict_assassment_plot(
  new_trial_analysis,
  saf_topic,
  select_btrt,
  select_analysis
)
}
\arguments{
\item{saf_topic}{Selected safety topic to analyze/the adverse event of interest}

\item{select_btrt}{Selected backgroundtreatments}

\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}
}
\description{
Title
}
