% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{nta_data_conflict_assassment_plot}
\alias{nta_data_conflict_assassment_plot}
\title{Graphic assessment of Prior Data conflict}
\usage{
nta_data_conflict_assassment_plot(
  new_trial_analysis,
  saf_topic,
  select_btrt,
  select_analysis
)
}
\arguments{
\item{new_trial_analysis}{Coordinates from the nta tab}

\item{saf_topic}{Selected safety topic to analyze/the adverse event of interest #nolint}

\item{select_btrt}{Selected backgroundtreatments}

\item{select_analysis}{Incidence proportion or Exposure-adjusted AE rate}
}
\description{
Prior data conflict assessment - compare prior, likelihood, and posterior in the NTA Tab #nolint
}
